Add: Class Filme and Media + Refactor: Class Serie
