+++ codigo/src/main/java/Filme.java
@@ -1,0 +1,25 @@
+import java.sql.Date;
+
+public class Filme extends Media {
+
+    private Integer duracao;
+    private Date dataLancamento;
+
+    public Filme(String nome, String genero, String idioma, Integer duracao, Date dataLancamento) {
+        super(nome, genero, idioma);
+        this.dataLancamento = dataLancamento;
+        this.duracao = duracao;
+    }
+
+    public Integer getDuracao() {
+        return duracao;
+    }
+
+    public Date getDataLancamento() {
+        return dataLancamento;
+    }
+
+    public Integer getDuracaoSegundos() {
+        return duracao * 60;
+    }
+}
